ref(onboarding): Extract the agent capability hovercard
The hovercard body was the longest thing in the setup card while saying
nothing about setup: it describes what the agent does and what it reports. Move
it, its capability list, and its trigger into AgentInfo.

Also drops Pi and OpenCode from the works-with line — neither is supported yet.
